Titanfall Dev Team
Titanfall 3 when?

Implemented a checkpoint system and respawning if you fall off of the map. Also added some more platforms so we have something to show for the playtest.

A few things unfortunately also broke in this version.

Bug 1: Collision for certain platforms will randomly not activate

Bug 2: Lane changing will break, causing the foreground to be considered the background

I think there's also a third bug, but I'll comment on it when I next work on this branch.